dom/bindings/test/test_bug1808352.html

<!doctype html>
<meta charset=utf-8>
<title>Test for Bug 1808352</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<div id=log></div>
<div id="testnode"></div>
<script>
function loadedFrame() {
    var node = document.getElementById("testnode");
    var frame = document.getElementById("frame");
    frame.contentWindow.document.adoptNode(node);
    document.adoptNode(node);
    var res;
    // Use a loop to ensure the JITs optimize the `id` getter access.
    for (var i = 0; i < 10_000; i++) {
        res = node.id;
    }
    is(res, "testnode", "expected node id");
    SimpleTest.finish();
}
SimpleTest.waitForExplicitFinish();
document.domain = "mochi.test"; // We want a cross-compartment adoptNode.
</script>
<iframe id="frame" src="http://test1.mochi.test:8888/tests/dom/bindings/test/file_bug1808352_frame.html"></iframe>
